Repackaging and javadoc
