+++ platforms/software/dependency-management/src/main/java/org/gradle/api/internal/artifacts/DefaultDependencyManagementServices.java
@@ -123,7 +123,7 @@ import org.gradle.internal.component.ResolutionFailureHandler;
 import org.gradle.internal.component.external.model.JavaEcosystemVariantDerivationStrategy;
 import org.gradle.internal.component.external.model.ModuleComponentArtifactMetadata;
 import org.gradle.internal.component.model.GraphVariantSelector;
-import org.gradle.internal.component.resolution.failure.FailureDescriberRegistry;
+import org.gradle.internal.component.resolution.failure.ResolutionFailureDescriberRegistry;
 import org.gradle.internal.event.ListenerManager;
 import org.gradle.internal.execution.ExecutionEngine;
 import org.gradle.internal.execution.InputFingerprinter;
@@ -491,7 +491,7 @@ public class DefaultDependencyManagementServices implements DependencyManagement
         }
 
         ResolutionFailureHandler createResolutionFailureProcessor(InstantiatorFactory instantiatorFactory, DocumentationRegistry documentationRegistry) {
-            FailureDescriberRegistry failureDescriberRegistry = FailureDescriberRegistry.standardRegistry(instantiatorFactory, documentationRegistry);
+            ResolutionFailureDescriberRegistry failureDescriberRegistry = ResolutionFailureDescriberRegistry.standardRegistry(instantiatorFactory, documentationRegistry);
             return new ResolutionFailureHandler(failureDescriberRegistry, documentationRegistry);
         }
