+++ platforms/software/dependency-management/src/main/java/org/gradle/internal/component/resolution/failure/describer/AmbiguousArtifactVariantsFailureDescriber.java
@@ -21,7 +21,7 @@ import org.gradle.api.internal.attributes.AttributeDescriber;
 import org.gradle.internal.component.AmbiguousArtifactVariantsException;
 import org.gradle.internal.component.model.AttributeDescriberSelector;
 import org.gradle.internal.component.resolution.failure.ResolutionCandidateAssessor.AssessedCandidate;
-import org.gradle.internal.component.resolution.failure.failuretype.AmbiguousResolutionFailure;
+import org.gradle.internal.component.resolution.failure.type.AmbiguousResolutionFailure;
 import org.gradle.internal.logging.text.TreeFormatter;
 
 import javax.inject.Inject;
