+++ platforms/software/dependency-management/src/main/java/org/gradle/internal/component/resolution/failure/type/ResolutionFailure.java
@@ -14,5 +14,24 @@
  * limitations under the License.
  */
 
-@org.gradle.api.NonNullApi
-package org.gradle.internal.component.resolution.failure.failuretype;
+package org.gradle.internal.component.resolution.failure.type;
+
+import org.gradle.api.internal.attributes.AttributesSchemaInternal;
+
+public class ResolutionFailure {
+    private final AttributesSchemaInternal schema;
+    private final String requestedName;
+
+    public ResolutionFailure(AttributesSchemaInternal schema, String requestedName) {
+        this.schema = schema;
+        this.requestedName = requestedName;
+    }
+
+    public AttributesSchemaInternal getSchema() {
+        return schema;
+    }
+
+    public String getRequestedName() {
+        return requestedName;
+    }
+}
